% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_iris.R
\docType{data}
\name{my_iris}
\alias{my_iris}
\title{Edgar Anderson's Iris Data}
\format{a data frame with 150 observations and 5 variables:
\describe{
\item{Sepal.Length}{sepal length, in cm}
\item{Sepal.Width}{sepal width, in cm}
\item{Petal.Length}{petal length, in cm}
\item{Petal.Width}{petal width, in cm}
\item{Species}{species of iris}
}}
\source{
Fisher, R. A. (1936) The use of multiple measurements in taxonomic problems.
\emph{Annals of Eugenics}, 7, Part II, 179–188.

The data were collected by Anderson, Edgar (1935). The irises of the Gaspe
Peninsula, \emph{Bulletin of the American Iris Society}, 59, 2–5.
}
\usage{
my_iris
}
\description{
The famous \code{\link[datasets]{iris}} dataset taken from the
\code{datasets} package to practice adding datasets to our R package.
}
\keyword{datasets}
