% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame}

\item{cl}{true class value of your training data}

\item{k_nn}{integer representing the number of neighbors}

\item{k_cv}{integer representing the number of folds}
}
\value{
A list with a vector of the predicted class for all observations and
 a numeric with the cross-validation misclassification error
}
\description{
This function is to to predict the output class using covariates.
}
\examples{
my_knn_cv(train = iris[, -5], cl = iris$Species, k_nn = 1, k_cv = 5)
my_knn_cv(train = iris[, -5], cl = iris$Species, k_nn = 5, k_cv = 5)

}
\keyword{prediction}
