% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting a linear model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula class object, similar to lm()}

\item{data}{input data frame}
}
\value{
a table with rows for each coefficient and columns
 for the Estimate, Std. Error, t value, and Pr(>|t|)
}
\description{
my_lm fits a linear model
}
\examples{
my_lm(mpg ~ hp + wt, mtcars)

}
