% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{number of folds}
}
\value{
a list with a numeric with the cross-validation error
}
\description{
This function is to predict output using covariates.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
