% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{Sample t-test Function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data}

\item{alternative}{a character string specifying the alternative hypothesis}

\item{mu}{a number indicating the null hypothesis value of the mean}
}
\value{
A list with the following elements:
  test_stat: the numeric test statistic
  df: the degrees of freedom.
  alternative: the value of the parameter `alternative`
  p_val: the numeric p-value.
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t.test(1:5, 2, alt = "greater")

}
\keyword{inference}
